requirements.txt
streamlit
firebase-admin
PyMuPDF
openai
python-docx
python-dotenv
pandas
numpy
matplotlib
requests
scikit-learn
Pillow
watchdog
tqdm
email-validator
validators
openai
langchain
tiktoken
firebase-admin
python-dotenv
gdown
transformers
